requirements.txt
# Core LLM & AI
openai>=1.12.0  # DeepSeek API compatible
cognee>=0.1.15  # Memory & Knowledge Graph
langgraph>=0.0.26
langchain>=0.1.0
langchain-community>=0.0.20

# Data Processing & Analytics
pandas>=2.2.0
polars>=0.20.0
duckdb>=0.10.0
pyarrow>=15.0.0

# Scientific Computing
numpy>=1.26.0
scipy>=1.12.0
statsmodels>=0.14.0
linearmodels>=5.3
cvxpy>=1.4.0
pymc>=5.10.0
scikit-learn>=1.3.0

# Graph & Network
networkx>=3.2

# Performance
numba>=0.59.0

# Excel & Office
openpyxl>=3.1.0
xlsxwriter>=3.2.0
python-docx>=1.1.0

# PDF & Reporting
reportlab>=4.0.0
weasyprint>=60.0

# Visualization & Dashboards
plotly>=5.18.0
kaleido>=0.2.1

# Financial Data APIs
requests>=2.31.0
sec-edgar-downloader>=5.0.0
beautifulsoup4>=4.12.0
lxml>=5.1.0

# Document Processing
PyPDF2>=3.0.0
pdfminer.six>=20221105
python-pptx>=0.6.23

# Database & Storage
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9  # PostgreSQL with pgvector
redis>=5.0.0
pymongo>=4.6.0

# Vector Database
chromadb>=0.4.22
pgvector>=0.2.4

# Configuration & Environment
python-dotenv>=1.0.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
pyyaml>=6.0.0

# Web Scraping & HTTP
httpx>=0.26.0
aiohttp>=3.9.0
ratelimit>=2.2.1
ratelimiter>=1.2.0

# Data Validation & Serialization
marshmallow>=3.20.0
jsonschema>=4.21.0

# Logging & Monitoring
loguru>=0.7.0
prometheus-client>=0.19.0

# Production Dependencies
tenacity>=8.2.0  # Retry logic with exponential backoff
python-json-logger>=2.0.0  # Structured logging
sentry-sdk>=1.40.0  # Error tracking (optional)

# Task Orchestration
prefect>=2.14.0
celery>=5.3.0

# Utilities
tqdm>=4.66.0
click>=8.1.0
jinja2>=3.1.0
python-dateutil>=2.8.0
pytz>=2024.1

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=24.1.0
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.0
pre-commit>=3.6.0

# API Development (for future web interface)
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0

# Frontend UI
streamlit>=1.28.0
